split gpu tests
